move & setup transaction model to separate app
