+++ transaction/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from transaction.models import Transaction
+
+admin.site.register(Transaction)
