+++ transaction/serializers.py
@@ -1,0 +1,17 @@
+from rest_framework import serializers
+
+from transaction.models import Transaction
+
+
+class TransactionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Transaction
+        fields = (
+            "id",
+            "sender",
+            "receiver",
+            "transfer_amount",
+            "commission",
+            "status",
+            "timestamp",
+        )
