+++ transaction/views.py
@@ -1,0 +1,9 @@
+from rest_framework import viewsets
+
+from transaction.models import Transaction
+from transaction.serializers import TransactionSerializer
+
+
+class TransactionViewSet(viewsets.ModelViewSet):
+    queryset = Transaction.objects.all()
+    serializer_class = TransactionSerializer
