+++ wallet/views.py
@@ -1,14 +1,9 @@
 from rest_framework import viewsets
 
-from wallet.models import Transaction, Wallet
-from wallet.serializers import TransactionSerializer, WalletSerializer
+from wallet.models import Wallet
+from wallet.serializers import WalletSerializer
 
 
 class WalletViewSet(viewsets.ModelViewSet):
     queryset = Wallet.objects.all()
     serializer_class = WalletSerializer
-
-
-class TransactionViewSet(viewsets.ModelViewSet):
-    queryset = Transaction.objects.all()
-    serializer_class = TransactionSerializer
